+++ headers/mpmath.h
@@ -195,7 +195,7 @@ if(denom==0L) overflow=1; else {(out).x = divide((arg).x,denom,bitshift);\
     (out).y = (arg).y*(real)
 
 #define CMPLXrecip(arg,out)    \
-   { double denom; denom = sqr((arg).x) + sqr((arg).y);\
+   { LDBL denom; denom = sqr((arg).x) + sqr((arg).y);\
      if(denom==0.0) {(out).x = 1.0e10;(out).y = 1.0e10;}else\
     { (out).x =  (arg).x/denom;\
      (out).y = -(arg).y/denom;}}
